Propagate the tensor-core projection to the distillation trainers
`DistillationTrainer._chunk` runs the same upcast-then-fp32-GEMM twice, once for
the student and once for the teacher, and the experimental async distillation
trainer has a third copy. Same change, same reasoning: both operands are already
the model dtype, so upcasting them buys nothing and costs the tensor cores.

The `grad_logits @ w_chunk.float()` in utils.py is deliberately left alone --
there `grad_logits` really is fp32, so that upcast is load-bearing.

+++ trl/experimental/async_distillation/async_distillation_trainer.py
@@ -225,7 +225,8 @@ def _jsd_loss_chunk(
         are `torch.no_grad()` sums for this chunk only — callers accumulate them across chunks and reduce across ranks,
         exactly as the non-chunked path already did.
     """
-    logits = hidden_chunk.float() @ lm_head_weight.float().t()
+    # Same as the other chunked projections: project in the model dtype, upcast afterwards.
+    logits = (hidden_chunk @ lm_head_weight.t()).float()
     if lm_head_bias is not None:
         logits = logits + lm_head_bias.float()
     if logit_scale != 1.0:
