+++ trl/trainer/distillation_trainer.py
@@ -112,7 +112,10 @@ def _chunk(h_s, w_s, b_s, s_scale, s_softcap, h_t, w_t, b_t, t_scale, t_softcap,
     # the backward, never `(chunk, V)`. ZeRO-3 shards the `lm_head`, so gather it tightly around each projection.
     # `logit_scale` (Cohere) / `final_logit_softcapping` (Gemma) are applied per model to match its full forward.
     with maybe_gather_lm_head_ctx(w_s, b_s):
-        student_logits = h_s.float() @ w_s.float().t()
+        # Project in the model dtype and upcast only afterwards, as `"nll"` and `transformers`'
+        # `ForCausalLMLoss` do. Upcasting `h_s` and `w_s` first forces a non-tensor-core fp32 GEMM
+        # and materialises an fp32 copy of the whole `lm_head` weight, per chunk.
+        student_logits = (h_s @ w_s.t()).float()
         if b_s is not None:
             student_logits = student_logits + b_s.float()
     if s_scale != 1.0:
@@ -123,7 +126,7 @@ def _chunk(h_s, w_s, b_s, s_scale, s_softcap, h_t, w_t, b_t, t_scale, t_softcap,
     # and the teacher accumulates no gradients (the teacher params are not frozen by `prepare_model`). Everything
     # downstream inherits this since `teacher_logits` is already detached.
     with maybe_gather_lm_head_ctx(w_t, b_t), torch.no_grad():
-        teacher_logits = h_t.float() @ w_t.float().t()
+        teacher_logits = (h_t @ w_t.t()).float()
         if b_t is not None:
             teacher_logits = teacher_logits + b_t.float()
     if t_scale != 1.0:
